requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
dnspython==2.7.0
idna==3.10
mysqlclient==2.2.6
pillow==11.1.0
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
# 最小化依赖，只保留 frontend 功能需要的包
django-cors-headers==4.3.1
geopy==2.4.1
googlemaps==4.10.0
hashids==1.3.1